Project_05/05_Capstone/requirements.txt
alembic==1.4.2
appdirs==1.4.4
astroid==2.4.0
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
awscli==1.18.98
Babel==2.8.0
botocore==1.17.21
certifi==2020.4.5.1
click==7.1.2
colorama==0.4.3
distlib==0.3.0
docutils==0.15.2
ecdsa==0.15
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pipenv==2018.11.26
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2020.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
